+++ soundpool/ios/soundpool/Sources/soundpool/SoundpoolPlugin.swift
@@ -1,0 +1,8 @@
+import Flutter
+import UIKit
+
+public class SoundpoolPlugin: NSObject, FlutterPlugin {
+    public static func register(with registrar: FlutterPluginRegistrar) {
+        SwiftSoundpoolPlugin.register(with: registrar)
+    }
+}